+++ butler_online_api/src/cron.rs
@@ -1,5 +1,4 @@
 extern crate diesel;
-use diesel::{MysqlConnection, r2d2};
 use crate::database::initialize_db_pool;
 use crate::wiederkehrend::gemeinsame_buchung;
 use crate::wiederkehrend::buchung;
@@ -10,7 +9,6 @@ mod gemeinsame_buchungen;
 mod kategorien;
 mod partner;
 mod health;
-mod openidconnect_configuration;
 mod gemeinsame_dauerauftraege;
 mod result_dto;
 mod user;
@@ -19,7 +17,6 @@ mod core;
 mod database;
 mod wiederkehrend;
 
-type DbPool = r2d2::Pool<r2d2::ConnectionManager<MysqlConnection>>;
 
 fn main() {
     dotenvy::dotenv().ok();
