+++ butler_online_api/src/database.rs
@@ -1,6 +1,8 @@
 use diesel::{MysqlConnection, r2d2};
 
-use crate::DbPool;
+
+pub type DbPool = r2d2::Pool<r2d2::ConnectionManager<MysqlConnection>>;
+pub type DbError = Box<dyn std::error::Error + Send + Sync>;
 
 pub fn initialize_db_pool() -> DbPool {
     let conn_spec = std::env::var("DATABASE_URL").expect("DATABASE_URL should be set");
