+++ butler_online_api/src/kategorien/input_http.rs
@@ -1,14 +1,13 @@
-use diesel::{r2d2, MysqlConnection};
 use serde::Serialize;
 
 use crate::kategorien::output_db;
 use actix_web::{delete, error, get, post, web, HttpResponse, Responder};
 use uuid::Uuid;
+use crate::database::DbPool;
 
 use crate::kategorien::model::{Kategorie, NeueKategorie};
 use crate::user::model::User;
 
-type DbPool = r2d2::Pool<r2d2::ConnectionManager<MysqlConnection>>;
 
 #[derive(Debug, Clone, Serialize)]
 pub struct KategorieDto {
