+++ butler_online_api/src/partner/input_http.rs
@@ -1,13 +1,11 @@
-use diesel::{r2d2, MysqlConnection};
 use serde::{Deserialize, Serialize};
 
 use crate::partner::output_db;
 use actix_web::{delete, error, get, post, web, HttpResponse, Responder};
+use crate::database::DbPool;
 
 use crate::partner::model::{NeuerPartnerStatus, PartnerStatus};
 use crate::user::model::User;
-
-type DbPool = r2d2::Pool<r2d2::ConnectionManager<MysqlConnection>>;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct NeuerPartnerStatusDto {
