+++ butler_online_api/src/wiederkehrend/gemeinsame_buchung.rs
@@ -9,7 +9,7 @@ use crate::gemeinsame_dauerauftraege::model::GemeinsamerDauerauftrag;
 pub fn verarbeite_gemeinsame_dauerauftraege(mut connection: &mut PooledConnection<ConnectionManager<MysqlConnection>>) {
     let start = SystemTime::now();
     let mut anzahl_verarbeiteter_buchungen = 0;
-    let auftraege = gemeinsame_dauerauftraege::output_db::finde_alle_gemeinsame_dauerauftraege_without_user(&mut connection).unwrap();
+    let auftraege = gemeinsame_dauerauftraege::output_db_unsafe::finde_alle_gemeinsame_dauerauftraege_without_user(&mut connection).unwrap();
     for dauerauftrag in auftraege.iter() {
         anzahl_verarbeiteter_buchungen += verarbeite_gemeinsame_buchung_dauerauftrag(&mut connection, dauerauftrag);
     }
